refactor function into its own pkg
