+++ core/src/main/java/com/airbnb/aerosolve/core/function/Spline.java
@@ -1,4 +1,4 @@
-package com.airbnb.aerosolve.core.util;
+package com.airbnb.aerosolve.core.function;
 
 import com.airbnb.aerosolve.core.FunctionForm;
 import com.airbnb.aerosolve.core.ModelRecord;
