+++ core/src/main/java/com/airbnb/aerosolve/core/models/DecisionTreeModel.java
@@ -3,19 +3,13 @@ package com.airbnb.aerosolve.core.models;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.IOException;
-import java.io.Serializable;
 import java.lang.StringBuilder;
 import java.util.Map;
 import java.util.List;
-import java.util.HashMap;
 import java.util.ArrayList;
-import java.util.PriorityQueue;
-import java.util.AbstractMap;
 
 import com.airbnb.aerosolve.core.*;
-import com.airbnb.aerosolve.core.util.FloatVector;
 import com.airbnb.aerosolve.core.util.Util;
-import com.airbnb.aerosolve.core.util.Spline;
 import lombok.Getter;
 import lombok.Setter;
