+++ core/src/test/java/com/airbnb/aerosolve/core/models/AdditiveModelTest.java
@@ -4,9 +4,9 @@ import com.airbnb.aerosolve.core.FeatureVector;
 import com.airbnb.aerosolve.core.ModelHeader;
 import com.airbnb.aerosolve.core.ModelRecord;
 import com.airbnb.aerosolve.core.FunctionForm;
-import com.airbnb.aerosolve.core.util.Function;
-import com.airbnb.aerosolve.core.util.Linear;
-import com.airbnb.aerosolve.core.util.Spline;
+import com.airbnb.aerosolve.core.function.Function;
+import com.airbnb.aerosolve.core.function.Linear;
+import com.airbnb.aerosolve.core.function.Spline;
 import com.airbnb.aerosolve.core.util.Util;
 import com.google.common.base.Optional;
 import org.junit.Test;
