+++ core/src/test/java/com/airbnb/aerosolve/core/util/LinearTest.java
@@ -1,9 +1,9 @@
 package com.airbnb.aerosolve.core.util;
 
 import com.airbnb.aerosolve.core.ModelRecord;
+import com.airbnb.aerosolve.core.function.Linear;
 import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
 import java.util.*;
 import static org.junit.Assert.assertEquals;
